transfer fetch to pages/index.js
